# openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionOut": {
        "properties": {
          "advisory_cache_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Cache Key"
          },
          "advisory_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Class"
          },
          "advisory_confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Confidence"
          },
          "advisory_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Provider"
          },
          "advisory_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Rationale"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "decision_hash": {
            "title": "Decision Hash",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "decision_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Decision Reasons",
            "type": "array"
          },
          "effect_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect Action"
          },
          "effect_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect State"
          },
          "effective_class": {
            "title": "Effective Class",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "gate_checks": {
            "items": {
              "$ref": "#/components/schemas/GateCheckOut"
            },
            "title": "Gate Checks",
            "type": "array"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "known_reason": {
            "title": "Known Reason",
            "type": "boolean"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "next_eligible_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Eligible At"
          },
          "payment_id": {
            "title": "Payment Id",
            "type": "string"
          },
          "policy_fingerprint": {
            "title": "Policy Fingerprint",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reason_map_fingerprint": {
            "title": "Reason Map Fingerprint",
            "type": "string"
          },
          "review_required": {
            "title": "Review Required",
            "type": "boolean"
          },
          "selected_action": {
            "title": "Selected Action",
            "type": "string"
          },
          "triage_rationale": {
            "title": "Triage Rationale",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "event_id",
          "payment_id",
          "created_at",
          "reason",
          "effective_class",
          "advisory_class",
          "advisory_rationale",
          "advisory_confidence",
          "advisory_provider",
          "advisory_cache_key",
          "known_reason",
          "review_required",
          "triage_rationale",
          "allowed_actions",
          "selected_action",
          "decision_reasons",
          "next_eligible_at",
          "reason_map_fingerprint",
          "policy_version",
          "policy_fingerprint",
          "decision_hash",
          "amount_minor",
          "currency",
          "method",
          "idempotency_key",
          "effect_action",
          "effect_state",
          "gate_checks"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
      "GateCheckOut": {
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "passed",
          "explanation"
        ],
        "title": "GateCheckOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaygroundInput": {
        "additionalProperties": false,
        "properties": {
          "amount_minor": {
            "maximum": 100000000.0,
            "minimum": 1.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "method": {
            "default": "card",
            "enum": [
              "card",
              "upi",
              "netbanking"
            ],
            "title": "Method",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scenario": {
            "enum": [
              "gateway_timeout",
              "insufficient_funds",
              "card_expired",
              "risk_threshold",
              "processor_code_z91"
            ],
            "title": "Scenario",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "scenario",
          "amount_minor"
        ],
        "title": "PlaygroundInput",
        "type": "object"
      },
      "PlaygroundReceipt": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/DecisionOut"
          },
          "decision_count": {
            "title": "Decision Count",
            "type": "integer"
          },
          "duplicate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate"
          },
          "effect_count": {
            "title": "Effect Count",
            "type": "integer"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "ingress_verified": {
            "title": "Ingress Verified",
            "type": "boolean"
          },
          "ledger_entry_count": {
            "title": "Ledger Entry Count",
            "type": "integer"
          },
          "ledger_valid": {
            "title": "Ledger Valid",
            "type": "boolean"
          },
          "request": {
            "$ref": "#/components/schemas/PlaygroundInput"
          },
          "safety_mode": {
            "const": "dry_run",
            "default": "dry_run",
            "title": "Safety Mode",
            "type": "string"
          }
        },
        "required": [
          "request",
          "decision",
          "duplicate",
          "ingress_verified",
          "ledger_valid",
          "event_count",
          "decision_count",
          "effect_count",
          "ledger_entry_count",
          "elapsed_ms"
        ],
        "title": "PlaygroundReceipt",
        "type": "object"
      },
...
      "RunInput": {
        "additionalProperties": false,
        "properties": {
          "amount_minor": {
            "maximum": 1000000.0,
            "minimum": 100.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "method": {
            "default": "card",
            "enum": [
              "card",
              "upi",
              "netbanking"
            ],
            "title": "Method",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scenario": {
            "default": "gateway_timeout",
            "enum": [
              "gateway_timeout",
              "insufficient_funds",
              "card_expired",
              "risk_threshold",
              "processor_code_z91"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "source": {
            "enum": [
              "synthetic",
              "razorpay_test"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "source",
          "amount_minor"
        ],
        "title": "RunInput",
        "type": "object"
      },
      "RunOut": {
        "properties": {
          "advice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CloudResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "checkout_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Key Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "event_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Source"
          },
          "ledger_valid": {
            "title": "Ledger Valid",
            "type": "boolean"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "safety_mode": {
            "const": "test_only_dry_run_recovery",
            "default": "test_only_dry_run_recovery",
            "title": "Safety Mode",
            "type": "string"
          },
          "scenario": {
            "title": "Scenario",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "webhook_deliveries": {
            "title": "Webhook Deliveries",
            "type": "integer"
          },
          "webhook_received": {
            "title": "Webhook Received",
            "type": "boolean"
          }
        },
        "required": [
          "run_id",
          "source",
          "scenario",
          "amount_minor",
          "method",
          "stage",
          "created_at",
          "order_id",
          "payment_id",
          "event_source",
          "error_code",
          "checkout_key_id",
          "decision",
          "advice",
          "webhook_received",
          "webhook_deliveries",
          "ledger_valid"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic payment-recovery evidence and read-only operator views.",
    "title": "Salvage read-only API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/demo/v1/runs": {
      "post": {
        "operationId": "run_scenario_demo_v1_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "header",
            "name": "x-salvage-playground",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Salvage-Playground"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaygroundInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scenario",
        "tags": [
          "local synthetic playground"
        ]
      }
    },
    "/demo/v1/runs/{run_id}/receipt": {
      "get": {
        "operationId": "download_receipt_demo_v1_runs__run_id__receipt_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Receipt",
        "tags": [
          "local synthetic playground"
        ]
      }
    },
...
    "/simulator/v1/runs": {
      "post": {
        "operationId": "create_simulator_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "connected test simulator"
        ]
      }
    },
    "/simulator/v1/runs/{run_id}": {
      "get": {
        "operationId": "detail_simulator_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "connected test simulator"
        ]
      }
    },
    "/simulator/v1/runs/{run_id}/receipt": {
      "get": {
        "operationId": "receipt_simulator_v1_runs__run_id__receipt_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt",
        "tags": [
          "connected test simulator"
        ]
      }
    },
    "/simulator/v1/runs/{run_id}/sync": {
      "post": {
        "operationId": "sync_simulator_v1_runs__run_id__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync",
        "tags": [
          "connected test simulator"
        ]
      }
    },
...
  }
}